export template macro
